Working on navigation and GUI, use NewGUI for fixing GUI
