+++ src/Navigation.java
@@ -1,0 +1,10 @@
+import java.awt.*;
+
+public class Navigation {
+    String[] buttons = {"Home","Symmetric","Asymmetric"};
+    String selected = "Home";
+    public void navigate(){
+//        homeButtonSelected.setBackground(Color.BLACK);
+//        symmetricSelected.setBackground(new Color(192,27,54));
+    }
+}
